Add OnePlus style
As of right now, it's missing OnePlus' signature red 1s in the hour, but I'm holding off on that for now. At some point I may add that but it'll take a bit more effort (create 4 text views, split the time into each, and color each correctly? not tonight thank you)